Create RandomInput.sv
This code creates random height and weight input values to depict the BMI calculation and further mapping it to the resistance gauge values in a range of 1 to 10. The main motive of this code is to show an algorithm that can accept input and perform robust and accurate computation to give an integer output of the resistance value.